+++ public/style2.css
@@ -1,0 +1,40 @@
+
+
+h1{
+    font-size: 50px;
+    
+}
+
+.container{
+    width: 90%;
+    margin: auto;
+}
+
+.flex{ display: flex;}
+
+.justify-center{ justify-content: center; }
+
+.justify-between{ justify-content: space-between; }
+
+.algin-center{ align-items: center; }
+
+.text-primary{ color: var(--primary-color); }
+
+.button-primary{
+    display: inline-block;
+    padding: 1rem 2rem;
+    margin-top: 1.5rem;
+    background-color: burlywood;
+    text-decoration: none;
+    color: white;
+    box-shadow: 2px 2px 5px #ddd;
+    transition: 0.2s ease-in;
+    border: none;
+    font-size: 1rem;
+}
+
+.button-primary:hover{
+    background-color: var(--primary-color-bright);
+
+}
+
